backend/requirements.txt

fastapi==0.119.0
uvicorn[standard]==0.37.0
python-multipart==0.0.20 
python-dotenv==1.1.1 
pydantic==2.12.0
qdrant-client==1.15.1 
sentence-transformers==5.1.1
pypdf==6.1.1 
python-docx==1.2.0 
pandas==2.3.3
boto3==1.40.50
